refactor(codegen): split constructor.rs (503 → directory of 4 files) (#153)
constructor.rs at 503 lines was the largest Rule 4 soft-cap allowlisted
file. Splits into a `constructor/` directory by responsibility:

  constructor/mod.rs          (123)  ConstructorCtx + is_dependency_type
                                     + RustGenerator::convert_constructor
                                     dispatcher
  constructor/params.rs       (107)  ExtractedParams + extract_*_param
                                     for the primary new() constructor
  constructor/field_init.rs   (171)  FieldInitCtx + body-walking logic
                                     (super, this.x = y, optional/Arc/
                                     Mutex wrapping, fill_missing_fields)
  constructor/di.rs           (158)  build_di_constructor + helpers for
                                     the secondary new_di() shape

Each file is under 200 lines (well under the 400-line soft cap). All
internal helpers are `pub(super)` — no public surface change. The two
crate-public items (`ConstructorCtx` struct + `is_dependency_type`
function) stay in `mod.rs` so callers in `class/mod.rs` continue to
import them as before.

Behavioral parity verified by full nextest run (248/248 incl.
test_http_equivalence_rust_server, test_reference_nestjs_transpiles_
and_compiles, test_multi_module_build).

Pre-commit run with TYRUS_SKIP_COVERAGE=1 due to unrelated #178 (main
coverage at 69.88% post-#170 syn bump).

Closes part of #153 (constructor.rs portion). interface.rs portion of
#153 was already addressed by #176 (now merged) which split type-decl
emission and dropped interface.rs to 158 lines.
